refactor: move preloading assets
